aws/Flask CRUD App with Aurora Serverless and Docker/App/requirements.txt
awsebcli==3.14.4
blessed==1.15.0
boto3==1.11.0
botocore==1.14.0
cached-property==1.4.3
cement==2.8.2
certifi==2018.8.24
chardet==3.0.4
click==6.7
colorama==0.3.9
docker==3.5.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==1.0.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
MarkupSafe==1.1
pathspec==0.5.5
python-dateutil==2.7.3
python-dotenv==0.21.0
PyYAML==3.13
requests==2.18.4
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
urllib3==1.21.1
wcwidth==0.1.7
websocket-client==0.51.0
Werkzeug==0.14.1
